Move DOM package from services to utils
