+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Dialog/Dialog.js
@@ -5,7 +5,7 @@ import {observer} from 'mobx-react';
 import React, {Fragment} from 'react';
 import type {Node} from 'react';
 import {Portal} from 'react-portal';
-import {afterElementsRendered} from '../../services/DOM';
+import {afterElementsRendered} from '../../utils/DOM';
 import Backdrop from '../Backdrop';
 import Button from '../Button';
 import dialogStyles from './dialog.scss';
