+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/DOM/__mocks__/afterElementsRendered.js
