+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/DOM/index.js
