+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/DOM/tests/afterElementsRendered.test.js
